feat: Add auto-release workflow (#406)
Add a new workflow file, auto-release.yaml, to enable automatic releases on Renovate. The workflow is triggered either manually or on a schedule. It checks if a release should be created based on the latest commit and the user who made the commit. If the commit is made by the projectorigin-renovate[bot] user, a new release is created using the previous tag as a starting point. The new version is determined by incrementing the patch number of the previous version. The release artifacts are then built and published.

Refactor release-published.yaml

Make a minor change to the release-published.yaml workflow file. The change adds a workflow_dispatch event to allow manual triggering of the workflow in addition to the existing release event trigger.